refactor: Implement useUpdateRoles hook
